non-custodial lightning receive.
after someone requests a lightning invoice, we now also wait for the
invoice to be paid. this is completely at the mercy of the phone OS on
how long it will keep the app alive before it dies waiting and pray to
god that the user pays invoice in a few seconds. from my testing on
regtest (android), seems ok for upto 10 seconds